third_party/blink/web_tests/virtual/container-style-queries-range/external/wpt/css/css-conditional/container-queries/query-evaluation-style-expected.txt

This is a testharness.js-based test.
Found 34 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] style(5 >= calc(3 + 1))
  assert_equals: expected "true" but got "false"
[FAIL] style(1px <= 1em)
  assert_equals: expected "true" but got "false"
[FAIL] style(5 >= 3)
  assert_equals: expected "true" but got "false"
[FAIL] style(3px >= 3px)
  assert_equals: expected "true" but got "false"
[FAIL] style(1em > 1px)
  assert_equals: expected "true" but got "false"
[FAIL] style(3turn > 3deg)
  assert_equals: expected "true" but got "false"
[FAIL] style(3% >= 3%)
  assert_equals: expected "true" but got "false"
[FAIL] style(3s > 3ms)
  assert_equals: expected "true" but got "false"
[FAIL] style(3dppx > 96dpi)
  assert_equals: expected "true" but got "false"
[FAIL] style(3 <= --x)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x >= 3)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x <= --x)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x >= --y)
  assert_equals: expected "true" but got "false"
[FAIL] style(--length > 3px)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x > 3px)
  assert_equals: expected "true" but got "false"
[FAIL] style(--number >= 3)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x >= 1)
  assert_equals: expected "true" but got "false"
[FAIL] style(--percentage > 3%)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x > 3%)
  assert_equals: expected "true" but got "false"
[FAIL] style(--angle < 1turn)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x < 1turn)
  assert_equals: expected "true" but got "false"
[FAIL] style(--time <= 1000ms)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x <= 1000ms)
  assert_equals: expected "true" but got "false"
[FAIL] style(3dppx > --resolution)
  assert_equals: expected "true" but got "false"
[FAIL] style(3dppx > --x)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x >= calc(3px + 3px))
  assert_equals: expected "true" but got "false"
[FAIL] style(--x <= var(--x))
  assert_equals: expected "true" but got "false"
[FAIL] style(calc(var(--x) + 1) >= var(--y))
  assert_equals: expected "true" but got "false"
[FAIL] style(var(--x) >= --x)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x <= 10em)
  assert_equals: expected "true" but got "false"
[FAIL] style(--length <= 30px)
  assert_equals: expected "true" but got "false"
[FAIL] style(10px <= 10px < 11px)
  assert_equals: expected "true" but got "false"
[FAIL] style(3 < --x <= 5)
  assert_equals: expected "true" but got "false"
[FAIL] style(--x >= --y > --z)
  assert_equals: expected "true" but got "false"
Harness: the test ran to completion.
